🛠refactor : history main 화면과 연결
